+++ js/variables.js
@@ -1,0 +1,58 @@
+//комментарии
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+//имя комментатора
+const NAMES = [
+  'Даша',
+  'Таня',
+  'Коля',
+  'Вася',
+  'Петя',
+  'Федя'
+];
+
+//описание - название фото
+const PHOTOS_NAMES = [
+  'Пруд с высоты птичьего полета',
+  'Указатель на пляж',
+  'Прилив',
+  'Девушка на пляже',
+  'Блюдо азиатской кухни',
+  'Черный автомобиль',
+  'Клубничный дессерт',
+  'Морс в двух стаканах',
+  'Самолет над пляжем',
+  'Обувница',
+  'Песчаный пляж',
+  'Белый ауди',
+  'Салат по-азиатски',
+  'КотоРолл',
+  'Модные тапочки',
+  'Над облаками',
+  'Симфонический оркестр',
+  'Ретро автомобиль',
+  'Не спится',
+  'Пальмы',
+  'Салат с лаймом',
+  'Закат над морем',
+  'Краб',
+  'Концерт',
+  'Сафари на джипе'
+];
+
+const MIN_LIKES = 15; // минимальное количество лайков
+const MAX_LIKES = 200;  //максимальное количество лайков
+const PHOTO_POSTS = 25;  //количество фотографий
+const MAX_COMMENTS = 30; //максимальное количество комментариев
+const MIN_NUMBER_AVATARS = 1;  //минимальный номер фото аватара
+const MAX_NUMBER_AVATARS = 6;  //максимальный номер фото аватара
+
+export {MESSAGES, NAMES, PHOTOS_NAMES, MIN_LIKES, MAX_LIKES, PHOTO_POSTS, MAX_COMMENTS, MIN_NUMBER_AVATARS, MAX_NUMBER_AVATARS};
